Refactor to get length of Actives
