+++ src/App.js
@@ -4,6 +4,7 @@ import { Provider } from 'react-redux';
 import Layout from './components/Layout';
 import Home from './components/Home';
 import store from './redux/store';
+import DisplayActives from './components/stocksCategory/actives/displayActives';
 
 function App() {
   return (
@@ -13,6 +14,7 @@ function App() {
           <Routes>
             <Route path="/" element={<Layout />} />
             <Route index element={<Home />} />
+            <Route path="stocks/actives" element={<DisplayActives />} />
           </Routes>
         </Provider>
       </BrowserRouter>
