+++ src/components/stocksCategory/actives/getActiveStocks.js
@@ -1,5 +1,5 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
-import apikey, { apiUrl } from '../misc/apiKey';
+import apikey, { apiUrl } from '../../../misc/apiKey';
 
 const getActiveStocks = createAsyncThunk('stocks/actives', async () => {
   const response = await fetch(`${apiUrl}actives?apikey=${apikey}`);
